backend/python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "adb-shell>=0.4.4",
    "fastapi>=0.118.0",
    "inquirerpy>=0.3.4",
    "numpy>=2.2.6",
    "opencv-python-headless>=4.12.0.88",
    "pillow>=11.3.0",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
    "pytesseract>=0.3.13",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
    "uiautomator2>=3.4.2",
    "uvicorn>=0.37.0",
]
